some tests and reorganizing
