+++ test/test_singly_linked_list.hpp
@@ -1,18 +1,14 @@
 #pragma once
+#include <unity.h>
 #include <LinkedList.hpp>
 
-using TESTING_TYPE = int;
-constexpr TESTING_TYPE list_default_testing_value = 100;
-constexpr TESTING_TYPE list_2default_testing_value = 150;
-constexpr unsigned short max_list_size = 100;
-
-void linked_list_empty()
+void singly_linked_list_empty()
 {
     SinglyLinkedList<TESTING_TYPE> list;
     TEST_ASSERT_TRUE(list.begin() == list.end());
     TEST_ASSERT_EQUAL(list.size(), 0);
 }
-void linked_list_standard_initializer()
+void singly_linked_list_standard_initializer()
 {
     for (unsigned short num_of_elements = 0; num_of_elements < max_list_size; ++num_of_elements)
     {
@@ -25,7 +21,7 @@ void linked_list_standard_initializer()
     }
 }
 
-void linked_list_push_back()
+void singly_linked_list_push_back()
 {
     SinglyLinkedList<TESTING_TYPE> list;
     unsigned short prev_list_size = list.size();
@@ -43,7 +39,7 @@ void linked_list_push_back()
     }
 }
 
-void linked_list_push_front()
+void singly_linked_list_push_front()
 {
     SinglyLinkedList<TESTING_TYPE> list;
     unsigned short prev_list_size = list.size();
@@ -55,13 +51,13 @@ void linked_list_push_front()
     for (unsigned short i = 0; i < max_list_size; ++i)
     {
         prev_list_size = list.size();
-        list.push_back(list_default_testing_value);
+        list.push_front(list_default_testing_value);
         TEST_ASSERT_EQUAL(prev_list_size + 1, list.size());
         TEST_ASSERT_TRUE(list_default_testing_value == list[0]);
     }
 }
 
-void linked_list_insert()
+void singly_linked_list_insert()
 {
     SinglyLinkedList<TESTING_TYPE> list;
     unsigned short prev_list_size = list.size();
@@ -78,7 +74,7 @@ void linked_list_insert()
     }
 }
 
-void linked_list_clear()
+void singly_linked_list_clear()
 {
     SinglyLinkedList<TESTING_TYPE> list;
     for (unsigned short list_size = 0; list_size < max_list_size; ++list_size)
@@ -94,7 +90,7 @@ void linked_list_clear()
     }
 }
 
-void linked_list_pop_back()
+void singly_linked_list_pop_back()
 {
     SinglyLinkedList<TESTING_TYPE> list;
     for (unsigned short list_size = 0; list_size < max_list_size; ++list_size)
@@ -121,7 +117,7 @@ void linked_list_pop_back()
     }
 }
 
-void linked_list_pop_front()
+void singly_linked_list_pop_front()
 {
     SinglyLinkedList<TESTING_TYPE> list;
     for (unsigned short list_size = 0; list_size < max_list_size; ++list_size)
@@ -148,7 +144,7 @@ void linked_list_pop_front()
     }
 }
 
-void linked_list_erase()
+void singly_linked_list_erase()
 {
     SinglyLinkedList<TESTING_TYPE> list;
     for (unsigned short list_size = 0; list_size < max_list_size; ++list_size)
@@ -179,7 +175,7 @@ void linked_list_erase()
     }
 }
 
-void linked_list_length()
+void singly_linked_list_length()
 {
     for (unsigned short i = 0; i < max_list_size; ++i)
     {
@@ -206,7 +202,7 @@ void linked_list_length()
     }
 }
 
-void linked_list_change_values()
+void singly_linked_list_change_values()
 {
     SinglyLinkedList<TESTING_TYPE> list(max_list_size, list_default_testing_value);
     for (unsigned short i = 0; i < max_list_size; ++i)
